Remove old notifier.
