+++ crates/cli/Cargo.toml
@@ -23,7 +23,7 @@ tracing = { workspace = true }
 [dev-dependencies]
 moon_cache = { path = "../cache" }
 moon_config = { path = "../config" }
-moon_notifier = { path = "../../legacy/core/notifier" }
+moon_notifier = { path = "../notifier" }
 moon_task = { path = "../task" }
 moon_task_runner = { path = "../task-runner" }
 moon_test_utils = { path = "../../legacy/core/test-utils" }
