+++ legacy/core/action-pipeline/src/subscribers/mod.rs
@@ -1,2 +1,3 @@
 pub mod local_cache;
 pub mod moonbase;
+pub mod webhooks;
