--- legacy/core/notifier/Cargo.toml
@@ -1,23 +1,0 @@
-[package]
-name = "moon_notifier"
-version = "0.0.1"
-edition = "2021"
-publish = false
-
-[dependencies]
-moon_app_context = { path = "../../../crates/app-context" }
-moon_emitter = { path = "../emitter" }
-moon_logger = { path = "../logger" }
-moon_utils = { path = "../utils" }
-async-trait = { workspace = true }
-ci_env = { workspace = true }
-miette = { workspace = true }
-reqwest = { workspace = true }
-serde = { workspace = true }
-serde_json = { workspace = true }
-starbase_styles = { workspace = true }
-tokio = { workspace = true }
-uuid = { workspace = true }
-
-[lints]
-workspace = true
