--- legacy/core/notifier/src/lib.rs
@@ -1,3 +1,0 @@
-mod webhooks;
-
-pub use webhooks::*;
